Pull test up in the hierarchy, to test both class directories and jars
